refactor(training): move NetworkTrainer to training.trainer_base
Split the NetworkTrainer class out of hv_train_network.py into its own module
so the file name matches its contents. hv_train_network.py is now just the
HunyuanVideo-specific entry point (hv_setup_parser + main) plus re-exports
for backward compatibility with architecture-specific training scripts that
import NetworkTrainer / helpers from musubi_tuner.hv_train_network.

SS_METADATA_* constants move with NetworkTrainer (it uses them) and are
re-exported from hv_train_network.py.

Fix sys.path fallback for --network_module short names: __file__ now lives
under musubi_tuner/training/, so step one extra level up to keep
`--network_module networks.lora` resolving to musubi_tuner/networks/lora.py.

hv_train_network.py: 2094 -> 87 lines.
training/trainer_base.py: new, 2061 lines.
